requirements.txt
pandas
numpy
scikit-learn
requests
python-dotenv
joblib
matplotlib
openai
fastapi
uvicorn
streamlit
xgboost